test: split source oracle integration behavior
